requirements.txt
#dependencies for the model
dvc==2.10.2
tensorflow==2.9.1
matplotlib==3.5.2
pandas==1.3.5
tqdm==4.64.0
pyYAML==6.0
boto3==1.24.11
joblib==1.1.0


#local lib
-e .



